+++ frontend/app/components/inplace-edit/field-directives/edit-duration/edit-duration.directive.html
@@ -12,5 +12,5 @@
          ng-required="fieldController.isRequired"
          title="{{ fieldController.editTitle }}"
          ng-model="numberValue"
-         float-value />
+         transform-float-value />
 </div>
