+++ frontend/app/components/inplace-edit/field-templates/edit/float.html
@@ -12,5 +12,5 @@
          ng-required="fieldController.isRequired"
          title="{{ fieldController.editTitle }}"
          ng-model="field.value"
-         float-value />
+         transform-float-value />
 </div>
