+++ frontend/app/components/input/transformers/transform-date.directive.ts
@@ -26,21 +26,22 @@
 // See doc/COPYRIGHT.rdoc for more details.
 //++
 
-module.exports = function($filter, TimezoneService) {
+function transformDate() {
   return {
     restrict:'A',
-    require: 'ngModel',
+    require: '^ngModel',
     link: function(scope, element, attrs, ngModelController) {
-      ngModelController.$parsers.push(function(value) {
-        if (!isNaN(value)) {
-          var minutes = Number(moment.duration(value, 'hours').asMinutes().toFixed(2));
-          return moment.duration(minutes, 'minutes');
+      ngModelController.$parsers.push(function(data) {
+        if (data === '') {
+          return null;
+        } else {
+          return data;
         }
-      });
-
-      ngModelController.$formatters.push(function(value) {
-        return TimezoneService.toHours(value);
       });
     }
   };
 };
+
+angular
+  .module('openproject')
+  .directive('transformDateValue', transformDate);
