+++ frontend/app/components/input/transformers/transform-duration.directive.ts
@@ -26,18 +26,25 @@
 // See doc/COPYRIGHT.rdoc for more details.
 //++
 
-module.exports = function($locale) {
-	return function(input) {
-    var decimalSep = $locale.NUMBER_FORMATS.DECIMAL_SEP;
-    var groupSep = $locale.NUMBER_FORMATS.GROUP_SEP;
+function transformDuration(TimezoneService) {
+  return {
+    restrict:'A',
+    require: 'ngModel',
+    link: function(scope, element, attrs, ngModelController) {
+      ngModelController.$parsers.push(function(value) {
+        if (!isNaN(value)) {
+          var minutes = Number(moment.duration(value, 'hours').asMinutes().toFixed(2));
+          return moment.duration(minutes, 'minutes');
+        }
+      });
 
-    var ret = (input) ? input.toString()
-                             .trim()
-                             .replace(groupSep, '')
-                             .replace(decimalSep, '.')
-                      :
-                        null;
-
-    return parseFloat(ret, 10);
-	};
+      ngModelController.$formatters.push(function(value) {
+        return TimezoneService.toHours(value);
+      });
+    }
+  };
 };
+
+angular
+  .module('openproject')
+  .directive('transformDurationValue', transformDuration);
