+++ frontend/app/components/wp-edit/field-types/wp-edit-duration-field.directive.html
@@ -2,7 +2,7 @@
        class="wp-inline-edit--field"
        wp-edit-field-requirements="vm.field.schema"
        ng-model="vm.workPackage[vm.fieldName]"
-       duration-value
+       transform-duration-value
        ng-required="vm.field.required"
        ng-blur="vm.handleUserBlur()"
        focus="vm.shouldFocus()"
