+++ frontend/app/components/wp-edit/field-types/wp-edit-float-field.directive.html
@@ -4,7 +4,7 @@
        ng-model="vm.workPackage[vm.fieldName]"
        ng-required="vm.field.required"
        ng-blur="vm.handleUserBlur()"
-       float-value
+       transform-float-value
        focus="vm.shouldFocus()"
        focus-priority="vm.shouldFocus()"
        ng-attr-id="{{vm.htmlId}}" />
