+++ frontend/app/ui_components/index.js
@@ -101,9 +101,4 @@ angular.module('openproject.uiComponents')
   .filter('ancestorsExpanded', require('./filters/ancestors-expanded-filter'))
   .directive('highlightCol', [require('./highlight-col-directive')])
   .directive('confirmPopup', ['$window', require('./confirm-popup-directive')])
-  .directive('clickOnKeypress', [require('./click-on-keypress-directive')])
-  .directive('floatValue', ['$filter', require('./float-directive')])
-  .directive('durationValue', ['$filter', 'TimezoneService', require('./duration-directive')])
-  .directive('transformDateValue', [require('./date-directive')])
-  .filter('external2internalFloat', ['$locale', require('./external-2-internal-float-filter')])
-  .filter('internal2externalFloat', ['$locale', require('./internal-2-external-float-filter')]);
+  .directive('clickOnKeypress', [require('./click-on-keypress-directive')]);
